staj_bot/requirements.txt
python-telegram-bot
selenium
webdriver-manager
ollama
